pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "ymago"
version = "0.1.0"
description = "An advanced, asynchronous command-line toolkit and Python library for generative AI media."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "google-genai>=1.36.0",
    "pydantic>=2.0,<3.0",
    "rich>=14.1.0",
    "tomli>=2.2.1",
    "typer>=0.17.4",
    "types-aiofiles>=24.1.0.20250822",
]

[project.optional-dependencies]
dev = [
    "aresponses>=3.0.0",
    "mypy>=1.17.1",
    "ruff>=0.13.0",
    "pytest-asyncio>=1.1.0",
    "pytest-xdist>=3.8.0",
]
docs = [
    "sphinx>=8.0.0",
    "sphinx-rtd-theme>=3.0.0",
    "sphinx-autodoc-typehints>=2.5.0",
    "myst-parser>=3.0.0",
    "sphinxcontrib-mermaid>=1.0.0",
]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "pytest-mock>=3.14.0",
    "pytest-benchmark>=5.0.0",
    "hypothesis>=6.100.0",

]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/docs",
    "/tests",
    "/.gitignore",
    "/.pre-commit-config.yaml",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ymago"]



[project.scripts]
ymago = "ymago.cli:app"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"] # flake8, pyflakes, bugbear, isort
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = ["google.*", "tenacity.*"]
ignore_missing_imports = true
